All users can do everything within the service
During recent work to allow users to create projects we ran into significant complexity in the test suite. Depending on the users role AND organisation they were assigned to, we were showing and hiding different bits of content through view logic and authorisation checks.

Our feature tests were becoming big and hard to work with as they were setting up contexts with lots of different variations. It was hard to conceptualise what the service did, and who was doing it.

In order to make things better we made a plan of which this is the first step:

1. Allow everyone to do everything. Using the abstract role of administrator as a tool to remove the fund manager and delivery partner role. This would allow projects to be added in a simpler way with less complexity. Given we're in private beta and we will have more control over what select fund managers and delivery partners click on, we can
2. Remove the concept of roles and rely on the organisation the user belongs to
3. Add read/write roles later when we need them

This change also renames feature tests where the user role is no longer relevant to comply with our new naming directive of users_*
