+++ app/helpers/organisation_helper.rb
@@ -1,7 +1,5 @@
 module OrganisationHelper
   def organisation_page_back_link(current_user, params)
-    return nil if current_user.delivery_partner?
-
     if current_user_on_home_page?(current_user, params)
       nil
     else
