+++ app/models/user.rb
@@ -5,11 +5,9 @@ class User < ApplicationRecord
 
   enum role: {
     administrator: "administrator",
-    delivery_partner: "delivery_partner",
-    fund_manager: "fund_manager",
   }
 
-  attribute :role, :string, default: "delivery_partner"
+  attribute :role, :string, default: "administrator"
 
   FORM_FIELD_TRANSLATIONS = {
     organisation_id: :organisation,
