+++ app/policies/user_policy.rb
@@ -1,21 +1,21 @@
 class UserPolicy < ApplicationPolicy
   def index?
-    user.administrator? || user.fund_manager?
+    user.administrator?
   end
 
   def show?
-    user.administrator? || user.fund_manager?
+    user.administrator?
   end
 
   def create?
-    user.administrator? || user.fund_manager?
+    user.administrator?
   end
 
   def update?
-    user.administrator? || user.fund_manager?
+    user.administrator?
   end
 
   def destroy?
-    user.administrator? || user.fund_manager?
+    user.administrator?
   end
 end
