+++ spec/features/staff/users_can_view_organisations_spec.rb
@@ -7,48 +7,25 @@ RSpec.feature "Users can view organisations" do
     end
   end
 
-  context "when the user is a fund manager" do
-    scenario "organisation index page" do
-      organisation = create(:organisation)
-      authenticate!(user: create(:fund_manager))
+  scenario "all organisations can be viewed" do
+    organisation = create(:organisation)
+    authenticate!(user: create(:administrator))
 
-      visit organisations_path
+    visit organisations_path
 
-      expect(page).to have_content(I18n.t("page_title.organisation.index"))
-      expect(page).to have_content organisation.name
-    end
-
-    scenario "can go back to the previous page" do
-      user_fund_manager = create(:fund_manager)
-
-      authenticate!(user: user_fund_manager)
-
-      visit organisations_path
-
-      click_on I18n.t("generic.link.back")
-
-      expect(page).to have_current_path(organisation_path(user_fund_manager.organisation))
-    end
+    expect(page).to have_content(I18n.t("page_title.organisation.index"))
+    expect(page).to have_content organisation.name
   end
 
-  context "when the user is a delivery partner" do
-    scenario "they cannot visit organisation index page" do
-      organisation = create(:organisation)
-      authenticate!(user: create(:delivery_partner, organisation: organisation))
+  scenario "can go back to the previous page" do
+    user = create(:administrator)
 
-      visit organisations_path
+    authenticate!(user: user)
 
-      expect(page).to have_content(I18n.t("page_title.errors.not_authorised"))
-    end
+    visit organisations_path
 
-    scenario "cannot see the organisation they don't belong to" do
-      organisation_they_belong_to = create(:organisation)
-      another_organisation = create(:organisation)
-      authenticate!(user: create(:delivery_partner, organisation: organisation_they_belong_to))
+    click_on I18n.t("generic.link.back")
 
-      visit organisation_path(another_organisation)
-
-      expect(page).to have_content(I18n.t("page_title.errors.not_authorised"))
-    end
+    expect(page).to have_current_path(organisation_path(user.organisation))
   end
 end
